Refactoring methods to be more general purpose.
